INT-MX 'source-only' metadata embedded by source and reported by each node
In the INT-MX header, unlike other metadata, 'source-only' domain
specific metadata is embedded in the packet after "DS Instruction"
and "DS Flags". This 'source-only' metadata should be reported by
each node including the source, transit, and sink nodes.

In the Telemetry Report, the 'source-only' metadata may be reported
in a truncated packet containing the INT-MX header, in domain specific
extension data, or in "Variable Optional Domain Specific Metadata"
with the corresponding bit set in "DSMdbits".